requirements.txt
pytest
discord-py-interactions
python-decouple